backend/requirements.txt
python-dotenv>=1.0.0
python-telegram-bot>=20.0
pyautogui>=0.9.54
pygetwindow>=0.0.9
pyperclip>=1.8.2
pillow>=10.0.0
customtkinter>=5.2.0
packaging
